Fix problems with color recall
